// the-lab-monorepo/apps/retailer-app-desktop/lib/types.ts
export interface Variant {
  id: string;
  size: string;
  pack: string;
  pack_count: number;
  weight_per_unit: string;
  price: number;
}

export interface Product {
  id: string;
  name: string;
  description: string;
  nutrition?: string;
  image_url: string;
  supplier_id: string;
  supplier_name: string;
  supplier_category: string;
  category_id: string;
  category_name: string;
  sell_by_block: boolean;
  units_per_block: number;
  price: number;
  variants?: Variant[];
  available_stock?: number;
}